requirements.txt
bcrypt==4.3.0
cffi==1.17.1
contourpy==1.3.1
cryptography==44.0.2
cx_Oracle==8.3.0
cycler==0.12.1
dnspython==2.7.0
fonttools==4.56.0
kiwisolver==1.4.8
matplotlib==3.10.1
mysql-connector-python==9.2.0
narwhals==1.29.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
pillow==11.1.0
plotly==6.0.0
prometheus_client==0.21.1
pycparser==2.22
pymongo==4.11.2
PyNaCl==1.5.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
scp==0.15.0
seaborn==0.13.2
six==1.17.0
tzdata==2025.1
